Add transition-group component
First css-line in the IUI library. WOW!

Closes #19
